Reapply D124184, [DebugInfo][InstrRef] Add a size operand to DBG_PHI
This was applied in 50ad99e2411f0, reverted in 65f7b3c2ce77a8, the problem
was that fp80 X86 registers that were spilt to the stack aren't expected by
LiveDebugValues. It pre-allocates a position number for all register sizes
that can be spilt, and 80 bits isn't exactly common.

The solution is to scan the register classes to find any unrecognised
register sizes, adn pre-allocate those position numbers, avoiding a later
assertion.